No.2 langchain春夏片-全面提升/13课上代码/pyproject.toml
[project]
name = "keshang13"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10,<3.14"
dependencies = [
    "dotenv>=0.9.9",
    "langchain>=1.2.0",
    "langchain-core>=1.1.3",
    "langchain-deepseek>=1.0.1",
    "langchain-tavily>=0.2.15",
    "langgraph-checkpoint-sqlite>=3.0.1",
]
